requirements.txt
# dbt
dbt-core>=1.7.0
dbt-duckdb>=1.7.0

# Database
duckdb>=0.9.0

# OpenAI client (for copilot-api compatibility)
openai>=1.0.0

# Visualization
plotly>=5.18.0
pandas>=2.0.0

# Web frameworks
streamlit>=1.30.0
flask>=3.0.0

# Utilities
python-dotenv>=1.0.0
